requirements.txt
click==8.1.3
Flask==2.1.3
Flask-Admin==1.6.0
Flask-Login==0.5.0
Flask-Migrate-4.0.0
Flask-SQLAlchemy==2.5.1
greenlet==2.0.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.0.3
MarkupSafe==2.1.2
pip==23.0
setuptools==49.2.1
SQLAlchemy==1.4.46
typing_extensions==4.4.0
Werkzeug==2.0.0
zipp==3.12.0
Flask-COMBO-JSONAPI==1.0.0
WTForms==3.0.1
email-validator==1.3.1
ComboJSONAPI==1.0.5
apispec==2.0.2
PyYAML==5.3.1

